Fix android 6 permission handling
Previously when permission was granted, the scan didn't start, it had to be retriggered by the user. Now the scan starts right after the permission gets granted.

The problem was certainly that json args given to the execute method were cleaned up when execute returned. This made it impossible to build Intent after the user granted permission.